requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
python-Levenshtein==0.21.1
